+++ requisicoes.py
@@ -1,41 +1,9 @@
 from datetime import datetime, timedelta
+from tokens import *
 import requests
 import pandas as pd
-import re
 import json
 import time
-
-def obter_token_bling():
-    """
-    Obtém o token de acesso do Bling a partir do Google Sheet
-    """
-
-    # URL para obter o token de acesso
-    token_bling_url = 'https://docs.google.com/spreadsheets/d/1ts-h9O8MKb1r16GNSl6ebNQnSknRq_EHrwtttjPpICU/export?format=csv&gid=0'
-
-    try:
-        data = pd.read_csv(token_bling_url)
-        token_acesso = data.iloc[0, 1]
-        return token_acesso
-    except Exception as e:
-        print("Erro ao acessar o Token na planilha do Google", str(e))
-        token_acesso = None
-
-    # Verifica se o token é válido
-    if not token_acesso:
-        raise Exception("Token de acesso não obtido. Abortando operação.")
-    
-# Obtém os tokens e define os headers do Bling/Pipefy
-bling_token = obter_token_bling()
-bling_headers = {"Authorization": f"Bearer {bling_token}", "Content-Type": "application/json"}
-
-pipefy_url = 'https://api.pipefy.com/graphql'
-pipefy_token = 'eyJhbGciOiJIUzUxMiJ9.eyJpc3MiOiJQaXBlZnkiLCJpYXQiOjE3MzkyOTYyOTYsImp0aSI6IjFmMzMyZmIwLTVhYmQtNDBkZi1iODA1LTcxMWQyMmI4ZjZmNiIsInN1YiI6MzAxMzgwNDQ2LCJ1c2VyIjp7ImlkIjozMDEzODA0NDYsImVtYWlsIjoibWlndWVsLnJvZG9scGhvQGdtYWlsLmNvbSJ9fQ.Exo9n9XMDKIAi9pgd1ySQC72rVYO6rLrxFPruVsxvBBl93Bk0qsEDM5A-yhQ7ojaqjDU8yhNoUpjTCSw_qFg6w'
-
-pipefy_headers = {
-    "Authorization": f"Bearer {pipefy_token}",
-    "Content-Type": "application/json"
-}
 
 def obter_compras_bling(status):
     """
@@ -92,46 +60,6 @@ def obter_pedidos_por_status(status):
         n_pedidos_bling.append(str(n_pedido))
 
     return n_pedidos_bling
-
-
-def obter_campos_pipefy():
-    """
-    Obtém informações referentes aos campos usados nos cards de um determinado pipe
-    """
-
-    pipe_id = "306043381"
-
-    # Define quais campos serão extraídos dos cards
-    query = '''
-    {
-    pipe(id: %s){
-        start_form_fields{
-            id
-            label
-            type
-            options
-            description
-            is_multiple
-        }
-        phases{
-            id
-            name
-        fields{
-            id
-            label
-        }
-        }
-    }
-    }
-    ''' % pipe_id
-
-    # Requisição POST para obter os dos campos incluídos nos cards
-    response = requests.post(pipefy_url, json={"query": query}, headers=pipefy_headers)
-    
-    if response.status_code == 200:
-        return(response.json())
-    else:
-        print("Erro ao buscar campos do Pipe:", response.status_code, response.json())
 
 
 def obter_cards_fase(pipe):
